Finished jobs page integration
I finished the basic implementation of the jobs page. It uses Indeed's job search API to find jobs based on the title and location entered by the user. It will not work as is, it requires an ID for the API request. If you'd like to test it, let me know so I can send you mine. I just didn't want to post my ID to github directly.